Code Analysis and Resharper clean-up
* Moved LuaManager into its own project
* Moved TimeManager into its own project
* Move some more lookup tables into the Lookups.lua file
* Incorporated Data and Documents folders into the Solution
* Refactored the SetValues function on ObjectTemplate
* Converted quite a few arrays into ienumerables on data classes
* Made most lists on data classes have private setters

+++ Source/SmaugCS/Skills/Dodge.cs
@@ -30,7 +30,7 @@ namespace SmaugCS.Skills
             if (victim.IsNpc())
                 chances = 60.GetLowestOfTwoNumbers(2*victim.Level);
             else
-                chances = Macros.LEARNED(victim, (int) skill.ID)/
+                chances = (int)Macros.LEARNED(victim, (int) skill.ID)/
                           (gameManager ?? GameManager.Instance).SystemData.DodgeMod;
 
             if (chances != 0 && victim.CurrentMorph != null)
